SenseForge - A modular Edge AI and Data Science toolkit/pyproject.toml
[build-system]
requires=['setuptools','wheel']
build-backend='setuptools.build_meta'

[project]
name='sensorkit'
version='0.1.0'
dependencies=['numpy','pandas','matplotlib','scipy','scikit-learn','pyyaml','pytest']
[project.scripts]
sensorkit='sensorkit.cli:main'
